rag-mongodb/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-mongodb>=0.1.0
langchain-community>=0.2.0
pymongo>=4.7.0
openai>=1.30.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
pypdf>=4.0.0
unstructured>=0.14.0
